did som stuff i guess
