+++ .clang-format
@@ -11,4 +11,7 @@ SpaceAfterCStyleCast: true
 Cpp11BracedListStyle: true
 SortIncludes: false
 AlwaysBreakTemplateDeclarations: Yes
+DerivePointerAlignment: false
+PointerAlignment: Left
+CommentNamespace: false
