--- includes/fecs/core/component.h
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace FECS {
-namespace Core {
-	
-	class Component {};
-}
-}
